[IMP] web: convert mobile signature test to hoot
task-4028335
