+++ addons/web/static/tests/views/widgets/signature.test.js
@@ -1,4 +1,5 @@
 import { NameAndSignature } from "@web/core/signature/name_and_signature";
+import { SignatureWidget } from "@web/views/widgets/signature/signature";
 
 import {
     defineModels,
@@ -8,14 +9,19 @@ import {
     onRpc,
     patchWithCleanup,
     contains,
+    toggleActionMenu,
+    clickModalButton,
 } from "@web/../tests/web_test_helpers";
 import { beforeEach, test, expect } from "@odoo/hoot";
-import { click, waitFor } from "@odoo/hoot-dom";
+import { click, queryFirst, waitFor } from "@odoo/hoot-dom";
+import { animationFrame } from "@odoo/hoot-mock";
 
 class Partner extends models.Model {
     display_name = fields.Char();
     product_id = fields.Many2one({ string: "Product Name", relation: "product" });
     sign = fields.Binary({ string: "Signature" });
+    signature = fields.Char();
+
     _records = [
         {
             id: 1,
@@ -217,4 +223,59 @@ test.tags("mobile")("Signature widget: highlight option on mobile", async () => 
     await click(".o_widget_signature button.o_sign_button");
     await waitFor(".modal .modal-body");
     expect(".modal .modal-body a.o_web_sign_auto_button").toHaveCount(0);
+});
+
+test.tags("mobile")("Signature widget works inside of a dropdown", async () => {
+    patchWithCleanup(SignatureWidget.prototype, {
+        async onClickSignature() {
+            await super.onClickSignature(...arguments);
+            expect.step("onClickSignature");
+        },
+        async uploadSignature({ signatureImage }) {
+            await super.uploadSignature(...arguments);
+            expect.step("uploadSignature");
+        },
+    });
+
+    await mountView({
+        type: "form",
+        resModel: "partner",
+        resId: 1,
+        arch: `
+                <form>
+                    <header>
+                        <button string="Dummy"/>
+                        <widget name="signature" string="Sign" full_name="display_name"/>
+                    </header>
+                    <field name="display_name" />
+                </form>
+            `,
+    });
+
+    // change display_name to enable auto-sign feature
+    await contains(".o_field_widget[name=display_name] input").edit("test");
+
+    // open the signature dialog
+    await toggleActionMenu();
+    await contains(".o_widget_signature button.o_sign_button").click();
+    await waitFor(".modal .modal-body");
+
+    // use auto-sign feature, might take a while
+    await contains(".o_web_sign_auto_button").click();
+
+    expect(".modal-footer button.btn-primary").toHaveCount(1);
+
+    let maxDelay = 100;
+    while (queryFirst(".modal-footer button.btn-primary")["disabled"] && maxDelay > 0) {
+        await animationFrame();
+        maxDelay--;
+    }
+
+    expect(maxDelay).toBeGreaterThan(0, { message: "Timeout exceeded" });
+
+    // close the dialog and save the signature
+    await clickModalButton({ text: "Adopt & Sign" });
+
+    expect(".modal-dialog").toHaveCount(0, { message: "Should have no modal opened" });
+    expect.verifySteps(["onClickSignature", "uploadSignature"]);
 });
